Добавить новый функционал

+++ productsService/db/client.py
@@ -28,7 +28,6 @@ def execute_sql_query(query: str, args: Tuple[Any, ...] | None = None, must_comm
 
     records = cur.fetchall()
     return records
-
 def get_products_with_sql_request(request: str, args: Tuple[Any, ...] | None = None) -> list[ProductModel] | None:
     records = execute_sql_query(request, args)
 
@@ -38,8 +37,8 @@ def get_products_with_sql_request(request: str, args: Tuple[Any, ...] | None = N
     products = []
     for record in records:
         products.append(
-            ProductModel(id=record[0], brand=record[1], price=record[2], size=record[3], color=record[4],
-                         country=record[5]))
+            ProductModel(id=record[0], price=record[1], size=record[2], color=record[3],
+                         country=record[4], sku=record[7]))
 
     return products
 
@@ -49,12 +48,12 @@ def get_cards_with_sql_request(request: str, args: Tuple[Any, ...] | None = None
         cards_models = []
         if len(records) != 0:
             for record in records:
-                cards_models.append(ProductsCardModel(id=record[0], title=record[1], description=record[2]))
+                cards_models.append(ProductsCardModel(id=record[0], title=record[1], description=record[2], brand=record[3],))
         else:
             return None
 
         for card in cards_models:
-            products = get_products_with_sql_request("SELECT * FROM public.products WHERE card_id = %s", (card.id,))
+            products = get_products_with_sql_request("SELECT * FROM public.products WHERE card_id = %s AND is_reserved = false", (card.id,))
 
             card.products = products if products else []
             card.count = len(products) if products else 0
@@ -70,7 +69,7 @@ def product_card_by_id(id: int) -> list[Any]:
         return get_cards_with_sql_request("SELECT * FROM public.products_card WHERE id = %s", (id,))
 
 def create_new_card(card: ProductsCardInputModel) -> list[Any]:
-        records =  execute_sql_query("insert into public.products_card (title, description) values (%s, %s) RETURNING id;", (card.title, card.description,), True)
+        records =  execute_sql_query("insert into public.products_card (title, description, brand) values (%s, %s, %s) RETURNING id;", (card.title, card.description, card.brand), True)
         return records[0]
 
 def edit_card(card: ProductsCardInputEditModel):
@@ -94,19 +93,23 @@ def delete_card(card_id: int):
                       (card_id,), True)
     return
 
+
 def all_products() -> list[Any]:
-    return get_products_with_sql_request("SELECT * FROM public.products")
+    return get_products_with_sql_request("SELECT * FROM public.products WHERE is_reserved = false")
 
 
 def product_by_id(id: int) -> list[Any]:
-    return get_products_with_sql_request("SELECT * FROM public.products WHERE id = %s", (id,))
+    return get_products_with_sql_request("SELECT * FROM public.products WHERE id = %s AND is_reserved = false", (id,))
+
+def product_by_card_id_and_sku(card_id: int, sku: int) -> list[Any]:
+    return get_products_with_sql_request("SELECT * FROM public.products WHERE card_id = %s AND sku = %s", (card_id, sku))
 
 def create_new_product(product: ProductInputModel) -> list[Any]:
     try:
         records = execute_sql_query(
-            "insert into public.products (brand, price, size, color, country, card_id) values "
-            "(%s, %s, %s, %s, %s, %s) RETURNING id;",
-            (product.brand, product.price, product.size, product.color, product.country, product.card_id),
+            "insert into public.products (price, size, color, country, card_id, is_reserved, sku) values "
+            "(%s, %s, %s, %s, %s, %s, %s) RETURNING id;",
+            (product.price, product.size, product.color, product.country, product.card_id, False, product.sku),
             True)
         return records[0]
     except psycopg2.errors.ForeignKeyViolation as ex:
@@ -114,6 +117,11 @@ def create_new_product(product: ProductInputModel) -> list[Any]:
 
 
 def edit_product(product: ProductInputEditModel):
+
+    products = get_products_with_sql_request("SELECT * FROM public.products WHERE id = %s AND is_reserved = true", (product.id,))
+    if products is not None:
+        raise ValueError("Продукт в заказе, невозможно изменить")
+
     attributes = inspect.getmembers(ProductInputEditModel, lambda a: not (inspect.isroutine(a)))
     elems = [x[1].keys() for x in [a for a in attributes if not (a[0].startswith('__'))] if x[0] == "model_fields"][0]
     sets_query = "set "
@@ -133,6 +141,14 @@ def edit_product(product: ProductInputEditModel):
     return
 
 def delete_product(product_id: int):
+    products = get_products_with_sql_request("SELECT * FROM public.products WHERE id = %s AND is_reserved = true", (product_id,))
+    if len(products) != 0:
+        raise ValueError("Продукт в заказе, невозможно удалить")
+
     execute_sql_query("delete from public.products where id = %s",
                       (product_id,), True)
+    return
+
+def reserve_product(product_id: int):
+    execute_sql_query("update public.products set is_reserved = true where id = %s", (product_id,), True)
     return
